refactor(core): combine TestBed interface and implementation to avoid cycles
This commit combines TestBed interface and implementation to avoid cycles in the import graph. There are no functional or API changes.
